LC-487: Brought the starting and stopping of timers out of WorkerThread to WorkerPool instead
